+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Report pitch rather than absolute pitch in fall detection warning
+- envs: Add `UpkieNavigation` environment
 
 ## [10.0.0] - 2025-10-20